voice_bot/requirements.txt
discord.py>=2.3.2
webrtcvad>=2.0.10
numpy>=1.24.0
openai-whisper>=20231117
aiohttp>=3.9.1
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
PyYAML>=6.0.1
python-dotenv>=1.0.0 